+++ noodles-bcf/src/lib.rs
@@ -5,6 +5,7 @@
 #[cfg(feature = "async")]
 pub mod r#async;
 
+pub mod fs;
 pub mod io;
 pub mod record;